First commit.
It is the first, but I already laid out the basis of this plugin.

There are two main namespaces:
  - the compressor   (vimball#compressor)
  - the decompressor (vimball#decompressor)

These namespaces' methods are exposed through a public API:
  - the ex API:
    - VimballPackage
    - VimballExtract
    - VimballList
    - legacy:
      - MkVimball
      - UseVimball
      - RmVimball

  - the functional API:
    - vimball#compress(dirname, filename, overwrite?)
    - vimball#decompress(bufnr, outdir)
    - vimball#list(bufnr)

The legacy API does not have any functional public equivalent, though
you can look in vimball#legacy if you really want to use it.