Changed behavior of extracting APIs
By extracting APIs, I mean :VimballExtract and vimball#extract()

These commands used to accept an optional path which defined the home
relative to which the vimball was extracted ($home/pack/vimball/start/$archive).

Now, when the path is specified, it defines the destination directory
into which the vimball is extracted.

However, the archive is still
placed relative to the home when the path is not specified.

This is done so ':VimballExtract my-plugin/' extracts the vimball into
my-plugin/ but  ':VimballExtract' installs the plugin.