refactor(archiver): Major rewrite to improve reliability
This fixes definitely the issue with the archive being included in the
archive because I resolve every path.

Also the archive header and writing a file into the archive was
extracted to its own function to allow building archives without using
the recursive archiving function. This is mainly done for the :MkVimball
command which accepts a list of file to archive from a buffer.