fix(extractor): Move fnamemodify call
This allows the first file of the vimball to be top-level.

Before, it wasn't possible, because the parent directory of the file as
a relative path was obtained with fnamemodify(), which returned '.'.

Then, isdirectory(/path/to/root/.) returned false because the root
wasn't even created yet, and mkdir(/path/to/root/.) tried to create the
directory twice because /path/to/root was the directory, but
/path/to/root/. was also the directory

Moving the fnamemodify call to the absolute directory fixes this bug and
should bring more reliability.