docs: Documented all of the remaining entries
This was quite long, but I can finally rest knowing that this project is
a 100% done.

I remember back when I was writing the *code* (the extractor was the
best part because I had the brilliant idea to use an iterator (I use
iterator everywhere, but...)), I just couldn't stop working on this
project. However, when I had to write the documentation, I was... less
excited.

What a surprise...

But is is DONE! (and it didn't take *that* long (just an hour and a half))